entire code is reorganized
