+++ driver/cmake-build-debug/.ninja_log
@@ -1,5 +1,22 @@
 # ninja log v5
+2	2193	1675410654714914443	CMakeFiles/driver.dir/drivermain.cpp.o	2f68432ec5ac55e7
+5	3974	1675409906087472946	CMakeFiles/driver.dir/libdriverutility.cpp.o	432a3eb83a90645d
 4518	4647	1675117728368774008	program	ba605559110ffc9e
+3	2248	1675410627351079011	CMakeFiles/driver.dir/drivercommapi.cpp.o	47d3bc0d574fdda3
+3	6309	1675409908419493295	CMakeFiles/driver.dir/libdriver.cpp.o	a7dcb16844560acf
+2193	2370	1675410654886913458	driver	b14451a2743b6bbf
 2	4518	1675117728240772315	CMakeFiles/program.dir/DriverLibrary.cpp.o	7aea53e6ff7e46db
 3	2848	1675117726560750086	CMakeFiles/program.dir/AuxiliaryTools.cpp.o	616deb096933ceb6
 2	3113	1675117726836753742	CMakeFiles/program.dir/main.cpp.o	5c225d646cbf0730
+6	4960	1675412194357244867	CMakeFiles/driver.dir/libdriverutility.cpp.o	432a3eb83a90645d
+3	5002	1675412194397244985	CMakeFiles/driver.dir/drivermain.cpp.o	2f68432ec5ac55e7
+5	5380	1675412194781246127	CMakeFiles/driver.dir/drivercommapi.cpp.o	47d3bc0d574fdda3
+4	7164	1675412196561235271	CMakeFiles/driver.dir/libdriver.cpp.o	a7dcb16844560acf
+7164	7315	1675412196713234151	driver	b14451a2743b6bbf
+10	7415	1675417785518596813	CMakeFiles/driver.dir/libdriverutility.cpp.o	432a3eb83a90645d
+4	7573	1675417785662597532	CMakeFiles/driver.dir/drivermain.cpp.o	2f68432ec5ac55e7
+9	8306	1675417786410601271	CMakeFiles/driver.dir/drivercommapi.cpp.o	47d3bc0d574fdda3
+6	10982	1675417789086614640	CMakeFiles/driver.dir/libdriver.cpp.o	a7dcb16844560acf
+10982	11329	1675417789430616359	driver	b14451a2743b6bbf
+2	1897	1675417811586726885	CMakeFiles/driver.dir/drivermain.cpp.o	2f68432ec5ac55e7
+1897	2060	1675417811754727722	driver	b14451a2743b6bbf
