+++ driver/cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Feb 02 11:22 +03
+Start testing: Feb 03 12:50 +03
 ----------------------------------------------------------
-End testing: Feb 02 11:22 +03
+End testing: Feb 03 12:50 +03
