+++ driver/cmake-build-debug/build.ninja
@@ -41,47 +41,53 @@ include CMakeFiles/rules.ninja
 
 cmake_ninja_workdir = /home/onur/CLionProjects/bbx15-driver/driver/cmake-build-debug/
 # =============================================================================
-# Object build statements for EXECUTABLE target program
+# Object build statements for EXECUTABLE target driver
 
 
 #############################################
-# Order-only phony target for program
+# Order-only phony target for driver
 
-build cmake_object_order_depends_target_program: phony || CMakeFiles/program.dir
+build cmake_object_order_depends_target_driver: phony || CMakeFiles/driver.dir
 
-build CMakeFiles/program.dir/main.cpp.o: CXX_COMPILER__program_Debug /home/onur/CLionProjects/bbx15-driver/driver/main.cpp || cmake_object_order_depends_target_program
-  DEP_FILE = CMakeFiles/program.dir/main.cpp.o.d
+build CMakeFiles/driver.dir/drivermain.cpp.o: CXX_COMPILER__driver_Debug /home/onur/CLionProjects/bbx15-driver/driver/drivermain.cpp || cmake_object_order_depends_target_driver
+  DEP_FILE = CMakeFiles/driver.dir/drivermain.cpp.o.d
   FLAGS = -g -fdiagnostics-color=always
-  OBJECT_DIR = CMakeFiles/program.dir
-  OBJECT_FILE_DIR = CMakeFiles/program.dir
+  OBJECT_DIR = CMakeFiles/driver.dir
+  OBJECT_FILE_DIR = CMakeFiles/driver.dir
 
-build CMakeFiles/program.dir/DriverLibrary.cpp.o: CXX_COMPILER__program_Debug /home/onur/CLionProjects/bbx15-driver/driver/DriverLibrary.cpp || cmake_object_order_depends_target_program
-  DEP_FILE = CMakeFiles/program.dir/DriverLibrary.cpp.o.d
+build CMakeFiles/driver.dir/libdriver.cpp.o: CXX_COMPILER__driver_Debug /home/onur/CLionProjects/bbx15-driver/driver/libdriver.cpp || cmake_object_order_depends_target_driver
+  DEP_FILE = CMakeFiles/driver.dir/libdriver.cpp.o.d
   FLAGS = -g -fdiagnostics-color=always
-  OBJECT_DIR = CMakeFiles/program.dir
-  OBJECT_FILE_DIR = CMakeFiles/program.dir
+  OBJECT_DIR = CMakeFiles/driver.dir
+  OBJECT_FILE_DIR = CMakeFiles/driver.dir
 
-build CMakeFiles/program.dir/AuxiliaryTools.cpp.o: CXX_COMPILER__program_Debug /home/onur/CLionProjects/bbx15-driver/driver/AuxiliaryTools.cpp || cmake_object_order_depends_target_program
-  DEP_FILE = CMakeFiles/program.dir/AuxiliaryTools.cpp.o.d
+build CMakeFiles/driver.dir/libdriverutility.cpp.o: CXX_COMPILER__driver_Debug /home/onur/CLionProjects/bbx15-driver/driver/libdriverutility.cpp || cmake_object_order_depends_target_driver
+  DEP_FILE = CMakeFiles/driver.dir/libdriverutility.cpp.o.d
   FLAGS = -g -fdiagnostics-color=always
-  OBJECT_DIR = CMakeFiles/program.dir
-  OBJECT_FILE_DIR = CMakeFiles/program.dir
+  OBJECT_DIR = CMakeFiles/driver.dir
+  OBJECT_FILE_DIR = CMakeFiles/driver.dir
+
+build CMakeFiles/driver.dir/drivercommapi.cpp.o: CXX_COMPILER__driver_Debug /home/onur/CLionProjects/bbx15-driver/driver/drivercommapi.cpp || cmake_object_order_depends_target_driver
+  DEP_FILE = CMakeFiles/driver.dir/drivercommapi.cpp.o.d
+  FLAGS = -g -fdiagnostics-color=always
+  OBJECT_DIR = CMakeFiles/driver.dir
+  OBJECT_FILE_DIR = CMakeFiles/driver.dir
 
 
 # =============================================================================
-# Link build statements for EXECUTABLE target program
+# Link build statements for EXECUTABLE target driver
 
 
 #############################################
-# Link the executable program
+# Link the executable driver
 
-build program: CXX_EXECUTABLE_LINKER__program_Debug CMakeFiles/program.dir/main.cpp.o CMakeFiles/program.dir/DriverLibrary.cpp.o CMakeFiles/program.dir/AuxiliaryTools.cpp.o
+build driver: CXX_EXECUTABLE_LINKER__driver_Debug CMakeFiles/driver.dir/drivermain.cpp.o CMakeFiles/driver.dir/libdriver.cpp.o CMakeFiles/driver.dir/libdriverutility.cpp.o CMakeFiles/driver.dir/drivercommapi.cpp.o
   FLAGS = -g
-  OBJECT_DIR = CMakeFiles/program.dir
+  OBJECT_DIR = CMakeFiles/driver.dir
   POST_BUILD = :
   PRE_LINK = :
-  TARGET_FILE = program
-  TARGET_PDB = program.dbg
+  TARGET_FILE = driver
+  TARGET_PDB = driver.dbg
 
 
 #############################################
@@ -117,7 +123,7 @@ build rebuild_cache: phony CMakeFiles/rebuild_cache.util
 #############################################
 # Folder: /home/onur/CLionProjects/bbx15-driver/driver/cmake-build-debug
 
-build all: phony program
+build all: phony driver
 
 # =============================================================================
 # Built-in targets
