+++ driver/cmake-build-debug/communication-register
@@ -1,1 +1,7 @@
 Tue Jan 31 01:28:49 2023	      slave writing: Invalid command
+Fri Feb  3 09:27:55 2023	      slave writing: Invalid command
+Fri Feb  3 10:44:07 2023	      slave writing: Invalid command
+Fri Feb  3 10:50:56 2023	      slave writing: Invalid command
+Fri Feb  3 11:16:38 2023	      slave writing: Invalid command
+Fri Feb  3 12:49:51 2023	      slave writing: Invalid command
+Fri Feb  3 12:50:13 2023	      slave writing: Invalid command
