+++ tester/CMakeLists.txt
@@ -3,4 +3,4 @@ project(tester)
 
 set(CMAKE_CXX_STANDARD 17)
 
-add_executable(tester main.cpp)
+add_executable(tester testermain.cpp)
